requirements.txt
django
django-pandas
pandas
plotly
django-import-export
